backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
psycopg[binary]==3.2.3
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.158.0
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
